Json Encode return on Autograde and  revamped code structure
