--- middleware/asdfas.py
@@ -1,3 +1,0 @@
-#!/usr/bin/python3
-def AddEm(x,y):
-	return x - y
