+++ middleware/util/autograde.go
@@ -1,19 +1,10 @@
-package pyrun
+package util
 
 import (
 	"fmt"
-	"io/ioutil"
-	"log"
-	"net/http"
-	"net/url"
 	"os"
 	"os/exec"
 )
-
-type UP struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-}
 
 type Grade struct {
 	Qid  string `json:"qid"`
@@ -25,30 +16,6 @@ type Ret struct {
 	Score string
 }
 
-// SendPOSTJSON -> sends a POST encoded with JSON
-func SendPostJSON(endpoint string, cd UP) string {
-	data := url.Values{
-		"username": {cd.Username},
-		"password": {cd.Password},
-	}
-
-	resp, err := http.PostForm(endpoint, data)
-	Check(err)
-
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	Check(err)
-
-	return string(body)
-}
-
-func Check(err error) {
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-}
-
 func CreatePyFile(code string, filename string) string {
 	// creates a python3 file
 
@@ -56,7 +23,7 @@ func CreatePyFile(code string, filename string) string {
 	file := fmt.Sprintf("%s.py", filename)
 	f, err := os.Create(file)
 	Check(err)
-	
+
 	_, err = f.Write([]byte("#!/usr/bin/python3\n"))
 	Check(err)
