+++ middleware/util/util.go
@@ -1,0 +1,12 @@
+package util
+
+import (
+	"log"
+)
+
+// Check ->
+func Check(err error) {
+	if err != nil {
+		log.Fatal(err.Error())
+	}
+}
